main: split control handling in a new control module
