+++ src/daemon/main.rs
@@ -1,4 +1,5 @@
 mod bitcoind;
+mod control;
 mod database;
 mod jsonrpc;
 mod revaultd;
@@ -6,34 +7,19 @@ mod threadmessages;
 
 use crate::{
     bitcoind::actions::{bitcoind_main_loop, start_bitcoind},
-    database::{
-        actions::setup_db,
-        interface::{db_deposits, db_tip, db_transactions, db_vault_by_deposit, RevaultTx},
-    },
+    control::handle_rpc_messages,
+    database::actions::setup_db,
     jsonrpc::{jsonrpcapi_loop, jsonrpcapi_setup},
-    revaultd::{BlockchainTip, RevaultD, VaultStatus},
-    threadmessages::*,
+    revaultd::RevaultD,
 };
 use common::config::Config;
-use revault_tx::{
-    bitcoin::Txid,
-    transactions::{
-        transaction_chain, CancelTransaction, EmergencyTransaction, RevaultTransaction,
-        UnvaultEmergencyTransaction, UnvaultTransaction,
-    },
-    txins::VaultTxIn,
-    txouts::VaultTxOut,
-};
 
 use std::{
     env,
     path::PathBuf,
     process,
     str::FromStr,
-    sync::{
-        mpsc::{self, Sender},
-        Arc, RwLock,
-    },
+    sync::{mpsc, Arc, RwLock},
     thread,
 };
 
@@ -51,28 +37,6 @@ fn parse_args(args: Vec<String>) -> Option<PathBuf> {
     }
 
     Some(PathBuf::from(args[2].to_owned()))
-}
-
-fn bitcoind_wallet_tx(
-    bitcoind_tx: &Sender<BitcoindMessageOut>,
-    txid: Txid,
-) -> Option<WalletTransaction> {
-    log::trace!("Sending WalletTx to bitcoind thread for {}", txid);
-
-    let (bitrep_tx, bitrep_rx) = mpsc::sync_channel(0);
-    bitcoind_tx
-        .send(BitcoindMessageOut::WalletTransaction(txid, bitrep_tx))
-        .unwrap_or_else(|e| {
-            log::error!("Sending 'wallettransaction' to bitcoind thread: {:?}", e);
-            process::exit(1);
-        });
-    bitrep_rx.recv().unwrap_or_else(|e| {
-        log::error!(
-            "Receiving 'wallettransaction' from bitcoind thread: {:?}",
-            e
-        );
-        process::exit(1);
-    })
 }
 
 fn daemon_main(mut revaultd: RevaultD) {
@@ -107,11 +71,11 @@ fn daemon_main(mut revaultd: RevaultD) {
     // The communication from us to the bitcoind thread
     let (bitcoind_tx, bitcoind_rx) = mpsc::channel();
 
-    let jsonrpc_thread = thread::spawn(move || {
+    let rpc_thread = thread::spawn(move || {
         jsonrpcapi_loop(rpc_tx, socket).unwrap_or_else(|e| {
             log::error!("Error in JSONRPC server event loop: {}", e.to_string());
             process::exit(1)
-        })
+        });
     });
 
     let revaultd = Arc::new(RwLock::new(revaultd));
@@ -120,441 +84,23 @@ fn daemon_main(mut revaultd: RevaultD) {
         bitcoind_main_loop(bitcoind_rx, bit_revaultd, &bitcoind).unwrap_or_else(|e| {
             log::error!("Error in bitcoind main loop: {}", e.to_string());
             process::exit(1)
-        })
+        });
     });
 
     log::info!(
         "revaultd started on network {}",
         revaultd.read().unwrap().bitcoind_config.network
     );
-    for message in rpc_rx {
-        match message {
-            RpcMessageIn::Shutdown => {
-                log::info!("Stopping revaultd.");
-                bitcoind_tx
-                    .send(BitcoindMessageOut::Shutdown)
-                    .unwrap_or_else(|e| {
-                        log::error!("Sending shutdown to bitcoind thread: {:?}", e);
-                        process::exit(1);
-                    });
-
-                jsonrpc_thread.join().unwrap_or_else(|e| {
-                    log::error!("Joining RPC server thread: {:?}", e);
-                    process::exit(1);
-                });
-                bitcoind_thread.join().unwrap_or_else(|e| {
-                    log::error!("Joining bitcoind thread: {:?}", e);
-                    process::exit(1);
-                });
-                process::exit(0);
-            }
-            RpcMessageIn::GetInfo(response_tx) => {
-                log::trace!("Got getinfo from RPC thread");
-
-                let (bitrep_tx, bitrep_rx) = mpsc::sync_channel(0);
-                bitcoind_tx
-                    .send(BitcoindMessageOut::SyncProgress(bitrep_tx))
-                    .unwrap_or_else(|e| {
-                        log::error!("Sending 'syncprogress' to bitcoind thread: {:?}", e);
-                        process::exit(1);
-                    });
-                let progress = bitrep_rx.recv().unwrap_or_else(|e| {
-                    log::error!("Receving 'syncprogress' from bitcoind thread: {:?}", e);
-                    process::exit(1);
-                });
-
-                // This means blockheight == 0 for IBD.
-                let BlockchainTip {
-                    height: blockheight,
-                    ..
-                } = db_tip(&db_path).unwrap_or_else(|e| {
-                    log::error!("Getting tip from db: {:?}", e);
-                    process::exit(1);
-                });
-
-                response_tx
-                    .send((network.to_string(), blockheight, progress))
-                    // TODO: a macro for the unwrap_or_else boilerplate..
-                    .unwrap_or_else(|e| {
-                        log::error!("Sending 'getinfo' result to RPC thread: {:?}", e);
-                        process::exit(1);
-                    });
-            }
-            RpcMessageIn::ListVaults((statuses, outpoints), response_tx) => {
-                log::trace!("Got listvaults from RPC thread");
-
-                let mut resp = Vec::<(u64, String, String, u32)>::new();
-                for (ref outpoint, ref vault) in revaultd.read().unwrap().vaults.iter() {
-                    if let Some(ref statuses) = statuses {
-                        if !statuses.contains(&vault.status) {
-                            continue;
-                        }
-                    }
-
-                    if let Some(ref outpoints) = &outpoints {
-                        if !outpoints.contains(&outpoint) {
-                            continue;
-                        }
-                    }
-
-                    resp.push((
-                        vault.txo.value,
-                        vault.status.to_string(),
-                        outpoint.txid.to_string(),
-                        outpoint.vout,
-                    ));
-                }
-
-                response_tx.send(resp).unwrap_or_else(|e| {
-                    log::error!("Sending 'listvaults' result to RPC thread: {}", e);
-                    process::exit(1);
-                });
-            }
-            RpcMessageIn::DepositAddr(response_tx) => {
-                log::trace!("Got 'depositaddr' request from RPC thread");
-                response_tx
-                    .send(revaultd.read().unwrap().deposit_address())
-                    .unwrap_or_else(|e| {
-                        log::error!("Sending 'depositaddr' result to RPC thread: {}", e);
-                        process::exit(1);
-                    });
-            }
-            RpcMessageIn::GetRevocationTxs(outpoint, response_tx) => {
-                log::trace!("Got 'getrevocationtxs' request from RPC thread");
-                let revaultd = revaultd.read().unwrap();
-                let xpub_ctx = revaultd.xpub_ctx();
-
-                // First, make sure the vault exists and is confirmed.
-                let vault = match revaultd.vaults.get(&outpoint) {
-                    None => None,
-                    Some(vault) => match vault.status {
-                        VaultStatus::Unconfirmed => None,
-                        _ => Some(vault),
-                    },
-                };
-                if let Some(vault) = vault {
-                    // Second, derive the fully-specified deposit txout. Note that we'd probably
-                    // store the index in the cache eventually, but until we get rid of this awful
-                    // mapping let's just use it.
-                    let index = revaultd
-                        .derivation_index_map
-                        .get(&vault.txo.script_pubkey)
-                        .unwrap_or_else(|| {
-                            log::error!("Unknown derivation index for: {:#?}", vault);
-                            process::exit(1);
-                        });
-                    let deposit_descriptor = revaultd.vault_descriptor.derive(*index);
-                    let vault_txin = VaultTxIn::new(
-                        outpoint,
-                        VaultTxOut::new(vault.txo.value, &deposit_descriptor, xpub_ctx),
-                    );
-
-                    // Third, re-derive all the transactions out of it.
-                    let unvault_descriptor = revaultd.unvault_descriptor.derive(*index);
-                    let cpfp_descriptor = revaultd.cpfp_descriptor.derive(*index);
-                    let emer_address = revaultd.emergency_address.clone();
-
-                    let (_, cancel, emergency, unvault_emer) = transaction_chain(
-                        vault_txin,
-                        &deposit_descriptor,
-                        &unvault_descriptor,
-                        &cpfp_descriptor,
-                        emer_address,
-                        xpub_ctx,
-                        revaultd.lock_time,
-                        revaultd.unvault_csv,
-                    )
-                    .unwrap_or_else(|e| {
-                        log::error!(
-                            "Deriving transactions for vault {:#?} (at '{}'): '{}'",
-                            vault,
-                            outpoint,
-                            e
-                        );
-                        process::exit(1);
-                    });
-
-                    response_tx
-                        .send(Some((cancel, emergency, unvault_emer)))
-                        .unwrap_or_else(|e| {
-                            log::error!("Sending 'getrevocationtxs' result to RPC thread: '{}'", e);
-                            process::exit(1);
-                        });
-                } else {
-                    response_tx.send(None).unwrap_or_else(|e| {
-                        log::error!(
-                            "Sending 'getrevocationtxs' (None) result to RPC thread: '{}'",
-                            e
-                        );
-                        process::exit(1);
-                    });
-                }
-            }
-            RpcMessageIn::RevocationTxs(
-                (outpoint, mut cancel, mut emer, mut unvault_emer),
-                response_tx,
-            ) => {
-                log::trace!("Got 'revocationtxs' from RPC thread");
-
-                let res = if revaultd.read().unwrap().vaults.get(&outpoint).is_some() {
-                    let _db_vault = db_vault_by_deposit(&db_path, &outpoint)
-                        .unwrap_or_else(|e| {
-                            log::error!("Getting vault from db: {}", e);
-                            process::exit(1);
-                        })
-                        .unwrap_or_else(|| {
-                            log::error!("(Insane db) None vault for '{}'", &outpoint);
-                            process::exit(1);
-                        });
-
-                    if cancel.finalize(&revaultd.read().unwrap().secp_ctx).is_err() {
-                        /* TODO: fetch from the SS */
-                    }
-                    if emer.finalize(&revaultd.read().unwrap().secp_ctx).is_err() {
-                        /* TODO: fetch from the SS */
-                    }
-                    if unvault_emer
-                        .finalize(&revaultd.read().unwrap().secp_ctx)
-                        .is_err()
-                    { /* TODO: fetch from the SS */ }
-
-                    // TODO: uncomment once above is implemented
-                    //db_store_revocation_txs(&revaultd.read().unwrap().db_file(), db_vault.id, cancel, emer, unvault_emer);
-                    None
-                } else {
-                    Some("Outpoint does not correspond to an existing vault".into())
-                };
-
-                response_tx.send(res).unwrap_or_else(|e| {
-                    log::error!("Sending 'revocationtxs' result to RPC thread: {}", e);
-                    process::exit(1);
-                });
-            }
-            RpcMessageIn::ListTransactions(outpoints, response_tx) => {
-                log::trace!("Got 'listtransactions' request from RPC thread");
-                let revaultd = revaultd.read().unwrap();
-                let xpub_ctx = revaultd.xpub_ctx();
-
-                // If they didn't provide us with a list of outpoints, catch'em all!
-                let outpoints = outpoints.unwrap_or_else(|| {
-                    db_deposits(&db_path)
-                        .unwrap_or_else(|e| {
-                            log::error!("Getting deposits from db: {}", e);
-                            process::exit(1);
-                        })
-                        .into_iter()
-                        .map(|db_vault| db_vault.deposit_outpoint)
-                        .collect()
-                });
-
-                let mut vaults = Vec::with_capacity(outpoints.len());
-                for outpoint in outpoints {
-                    if let Some(vault) = revaultd.vaults.get(&outpoint) {
-                        let db_vault = db_vault_by_deposit(&db_path, &outpoint)
-                            .unwrap_or_else(|e| {
-                                log::error!("Getting vault from db: {}", e);
-                                process::exit(1);
-                            })
-                            .unwrap_or_else(|| {
-                                log::error!("(Insane db) None vault for '{}'", &outpoint);
-                                process::exit(1);
-                            });
-                        let mut txs = db_transactions(&db_path, db_vault.id, &[])
-                            .unwrap_or_else(|e| {
-                                log::error!("Getting transactions (all) from db: {}", e);
-                                process::exit(1);
-                            })
-                            .into_iter();
-
-                        let deposit_tx = txs
-                            .find(|db_tx| matches!(db_tx.tx, RevaultTx::Deposit(_)))
-                            .map(|tx| assert_tx_type!(tx.tx, Deposit, "We just found it"))
-                            .unwrap_or_else(|| {
-                                log::error!("Vault without deposit tx in db for {}", outpoint);
-                                process::exit(1);
-                            });
-                        let wallet_tx = bitcoind_wallet_tx(&bitcoind_tx, deposit_tx.0.txid());
-                        let deposit = TransactionResource {
-                            wallet_tx,
-                            tx: deposit_tx,
-                            // The deposit is always signed, if we heard about it in
-                            // the first place
-                            is_signed: true,
-                        };
-
-                        // Get the descriptors in case we need to derive the transactions (not signed
-                        // yet, ie not in DB).
-                        // One day, we could try to be smarter wrt free derivation but it's not
-                        // a priority atm.
-                        let index = revaultd
-                            .derivation_index_map
-                            .get(&vault.txo.script_pubkey)
-                            .unwrap_or_else(|| {
-                                log::error!("Unknown derivation index for: {:#?}", vault);
-                                process::exit(1);
-                            });
-                        let deposit_descriptor = revaultd.vault_descriptor.derive(*index);
-                        let vault_txin = VaultTxIn::new(
-                            outpoint,
-                            VaultTxOut::new(vault.txo.value, &deposit_descriptor, xpub_ctx),
-                        );
-                        let unvault_descriptor = revaultd.unvault_descriptor.derive(*index);
-                        let cpfp_descriptor = revaultd.cpfp_descriptor.derive(*index);
-                        let emer_address = revaultd.emergency_address.clone();
-
-                        // We can always re-generate the Unvault out of the descriptor if it's
-                        // not in DB..
-                        let mut unvault_tx = txs
-                            .find(|db_tx| matches!(db_tx.tx, RevaultTx::Unvault(_)))
-                            .map(|tx| assert_tx_type!(tx.tx, Unvault, "We just found it"))
-                            .unwrap_or_else(|| {
-                                UnvaultTransaction::new(
-                                    vault_txin.clone(),
-                                    &unvault_descriptor,
-                                    &cpfp_descriptor,
-                                    xpub_ctx,
-                                    revaultd.lock_time,
-                                )
-                                .unwrap_or_else(|e| {
-                                    log::error!("Deriving unvault for '{}': {}", outpoint, e);
-                                    process::exit(1);
-                                })
-                            });
-                        let unvault_txin = unvault_tx
-                            .unvault_txin(&unvault_descriptor, xpub_ctx, revaultd.unvault_csv)
-                            .expect("Just created it");
-                        let wallet_tx = bitcoind_wallet_tx(
-                            &bitcoind_tx,
-                            unvault_tx.inner_tx().global.unsigned_tx.txid(),
-                        );
-                        // The transaction is signed if we did sign it, or if others did (eg
-                        // non-stakeholder managers) and we noticed it from broadcast.
-                        // TODO: maybe a is_finalizable upstream ? finalize() is pretty costy
-                        let is_signed =
-                            unvault_tx.finalize(&revaultd.secp_ctx).is_ok() || wallet_tx.is_some();
-                        let unvault = TransactionResource {
-                            wallet_tx,
-                            tx: unvault_tx,
-                            is_signed,
-                        };
-
-                        // .. But not the spend, as it's dynamically chosen by the managers and
-                        // could be anything!
-                        let spend = if let Some(mut tx) = txs
-                            .find(|db_tx| matches!(db_tx.tx, RevaultTx::Spend(_)))
-                            .map(|tx| assert_tx_type!(tx.tx, Spend, "We just found it"))
-                        {
-                            let wallet_tx = bitcoind_wallet_tx(
-                                &bitcoind_tx,
-                                tx.inner_tx().global.unsigned_tx.txid(),
-                            );
-                            // TODO: maybe a is_finalizable upstream ? finalize() is pretty costy
-                            let is_signed =
-                                tx.finalize(&revaultd.secp_ctx).is_ok() || wallet_tx.is_some();
-                            Some(TransactionResource {
-                                wallet_tx,
-                                tx,
-                                is_signed,
-                            })
-                        } else {
-                            None
-                        };
-
-                        // The cancel transaction is deterministic, so we can always return it.
-                        let mut cancel_tx = txs
-                            .find(|db_tx| matches!(db_tx.tx, RevaultTx::Cancel(_)))
-                            .map(|tx| assert_tx_type!(tx.tx, Cancel, "We just found it"))
-                            .unwrap_or_else(|| {
-                                CancelTransaction::new(
-                                    unvault_txin.clone(),
-                                    None,
-                                    &deposit_descriptor,
-                                    xpub_ctx,
-                                    revaultd.lock_time,
-                                )
-                            });
-                        let wallet_tx = bitcoind_wallet_tx(
-                            &bitcoind_tx,
-                            cancel_tx.inner_tx().global.unsigned_tx.txid(),
-                        );
-                        // TODO: maybe a is_finalizable upstream ? finalize() is pretty costy
-                        let is_signed =
-                            cancel_tx.finalize(&revaultd.secp_ctx).is_ok() || wallet_tx.is_some();
-                        let cancel = TransactionResource {
-                            wallet_tx,
-                            tx: cancel_tx,
-                            is_signed,
-                        };
-
-                        // The emergency transaction is deterministic, so we can always return it.
-                        let mut emergency_tx = txs
-                            .find(|db_tx| matches!(db_tx.tx, RevaultTx::Emergency(_)))
-                            .map(|tx| assert_tx_type!(tx.tx, Emergency, "We just found it"))
-                            .unwrap_or_else(|| {
-                                EmergencyTransaction::new(
-                                    vault_txin,
-                                    None,
-                                    emer_address,
-                                    revaultd.lock_time,
-                                )
-                            });
-                        let wallet_tx = bitcoind_wallet_tx(
-                            &bitcoind_tx,
-                            emergency_tx.inner_tx().global.unsigned_tx.txid(),
-                        );
-                        // TODO: maybe a is_finalizable upstream ? finalize() is pretty costy
-                        let is_signed = emergency_tx.finalize(&revaultd.secp_ctx).is_ok()
-                            || wallet_tx.is_some();
-                        let emergency = TransactionResource {
-                            wallet_tx,
-                            tx: emergency_tx,
-                            is_signed,
-                        };
-
-                        // Same for the second emergency.
-                        let mut unvault_emergency_tx = txs
-                            .find(|db_tx| matches!(db_tx.tx, RevaultTx::UnvaultEmergency(_)))
-                            .map(|tx| assert_tx_type!(tx.tx, UnvaultEmergency, "We just found it"))
-                            .unwrap_or_else(|| {
-                                UnvaultEmergencyTransaction::new(
-                                    unvault_txin,
-                                    None,
-                                    revaultd.emergency_address.clone(),
-                                    revaultd.lock_time,
-                                )
-                            });
-                        let wallet_tx = bitcoind_wallet_tx(
-                            &bitcoind_tx,
-                            unvault_emergency_tx.inner_tx().global.unsigned_tx.txid(),
-                        );
-                        // TODO: maybe a is_finalizable upstream ? finalize() is pretty costy
-                        let is_signed = unvault_emergency_tx.finalize(&revaultd.secp_ctx).is_ok()
-                            || wallet_tx.is_some();
-                        let unvault_emergency = TransactionResource {
-                            wallet_tx,
-                            tx: unvault_emergency_tx,
-                            is_signed,
-                        };
-
-                        vaults.push(VaultTransactions {
-                            outpoint,
-                            deposit,
-                            unvault,
-                            spend,
-                            cancel,
-                            emergency,
-                            unvault_emergency,
-                        });
-                    }
-                }
-
-                response_tx.send(vaults).unwrap_or_else(|e| {
-                    log::error!("Sending 'listtransactions' result to RPC thread: {}", e);
-                    process::exit(1);
-                });
-            }
-        }
-    }
+    // Handle RPC commands until we die.
+    handle_rpc_messages(
+        revaultd,
+        db_path,
+        network,
+        rpc_rx,
+        bitcoind_tx,
+        bitcoind_thread,
+        rpc_thread,
+    );
 }
 
 // This creates the log file automagically if it doesn't exist, and logs on stdout
